cleanner code. version 1.0

+++ main.py
@@ -3,26 +3,53 @@ import shutil
 
 downloads = "/home/abdelrahman/Downloads"
 music = "/home/abdelrahman/Music"
-Videos = "/home/abdelrahman/Videos"
-Pictures = "/home/abdelrahman/Pictures"
-Documents = "/home/abdelrahman/Documents"
-files = os.listdir(downloads)
+videos = "/home/abdelrahman/Videos"
+pictures = "/home/abdelrahman/Pictures"
+documents = "/home/abdelrahman/Documents"
 
-file_types = {Pictures: [".png", ".jpg", ".jpeg"]
-              , Documents: [".txt", ".pdf", ".docs", ".yaml"],
+file_types = {pictures: [".png", ".jpg", ".jpeg"]
+              , documents: [".txt", ".pdf", ".docs", ".yaml"],
                 music : [".wav", ".mp3"],
-                 Videos:[".mp4"]}
+                 videos:[".mp4"]}
 
-print(files)
+def access_files():
+    files = os.listdir(downloads)
 
-for file in files:
-    name, extension = os.path.splitext(file)
-    file_path = os.path.join(downloads, file)
-    print(extension)
+    for file in files:
+        file_path = os.path.join(downloads, file)
 
-    for category, extensions in file_types.items():
-        if extension.lower() in extensions:
-            shutil.move(file_path, category)
-           
+        if not os.path.isfile(file_path):
+            continue
+
+
+        name, extension = os.path.splitext(file)
+
+        for category, extensions in file_types.items():
+            if extension.lower() in extensions:
+                os.makedirs(category, exist_ok = True)
+                shutil.move(file_path, category)
+                break
+
+
+def logging():
+
+    return
+
+
+
+
+
+def main():
+    access_files()
+
+
+if __name__ == "__main__":
+    main()
+
+
+
+
+
+
